changed field name to fit conventions
moved maximum winds and recharge to ability model
