+++ CSharpSourceCode/Extensions/ExtendedInfoSystem/HeroExtendedInfo.cs
@@ -13,6 +13,7 @@ using TOR_Core.CampaignMechanics.Religion;
 using TOR_Core.CharacterDevelopment;
 using TOR_Core.CharacterDevelopment.CareerSystem;
 using TOR_Core.CampaignMechanics.CustomResources;
+using TOR_Core.Models;
 using TOR_Core.Utilities;
 
 namespace TOR_Core.Extensions.ExtendedInfoSystem
@@ -97,76 +98,11 @@ namespace TOR_Core.Extensions.ExtendedInfoSystem
             get
             {
                 if (!(Game.Current.GameType is Campaign)) return 50;
-                else
-                {
-                    if (BaseCharacter.HeroObject != null && BaseCharacter.HeroObject != Hero.MainHero && BaseCharacter.HeroObject.Occupation == Occupation.Lord && BaseCharacter.HeroObject.IsSpellCaster()) return 100f;
-                    ExplainedNumber explainedNumber = new ExplainedNumber(10f, false, null);
-                    SkillHelper.AddSkillBonusForCharacter(TORSkills.SpellCraft, TORSkillEffects.MaxWinds, BaseCharacter, ref explainedNumber);
-                    if (Hero.MainHero.HasAnyCareer())
-                    {
-                        if (BaseCharacter.HeroObject == Hero.MainHero)
-                        {
-                            CareerHelper.ApplyBasicCareerPassives(Hero.MainHero,ref  explainedNumber, PassiveEffectType.WindsOfMagic,false);
-                            
-                            if (CareerChoices.Contains("DarkVisionPassive4"))
-                            {
-                                var spellCount = Hero.MainHero.GetExtendedInfo().AcquiredAbilities.Count;
-                                var choice = TORCareerChoices.GetChoice("DarkVisionPassive4");
-                                explainedNumber.Add(choice.GetPassiveValue() * spellCount);
-                            }
-
-                            if (CareerChoices.Contains("DiscipleOfAccursedPassive4"))
-                            {
-                                var characterEquipment = Hero.MainHero.CharacterObject.GetCharacterEquipment();
-                                foreach (var item in characterEquipment)
-                                {
-                                    var choice = TORCareerChoices.GetChoice("DiscipleOfAccursedPassive4");
-                                    if (item.IsMagicalItem())
-                                    {
-                                        explainedNumber.Add(choice.GetPassiveValue());
-                                    } 
-                                }
-                            }
-                        }
-                        else if (BaseCharacter.HeroObject.PartyBelongedTo!=null && BaseCharacter.HeroObject.PartyBelongedTo.IsMainParty)
-                        {
-                            if (Hero.MainHero != null)
-                            {
-                                var choices = Hero.MainHero.GetAllCareerChoices();
-                                if (choices.Contains("EnvoyOfTheLadyPassive3"))
-                                {
-                                    var choice = TORCareerChoices.GetChoice("EnvoyOfTheLadyPassive3");
-                                    explainedNumber.Add(choice.GetPassiveValue());
-                                }
-                            
-                                if (choices.Contains("LieOfLadyPassive2"))
-                                {
-                                    var choice = TORCareerChoices.GetChoice("LieOfLadyPassive2");
-                                    explainedNumber.Add(choice.GetPassiveValue());
-                                }
-                                if (choices.Contains("LieOfLadyPassive2"))
-                                {
-                                    var choice = TORCareerChoices.GetChoice("LieOfLadyPassive2");
-                                    explainedNumber.Add(choice.GetPassiveValue());
-                                }
-                            
-                                if (choices.Contains("WellspringOfDharPassive3"))
-                                {
-                                    var choice = TORCareerChoices.GetChoice("WellspringOfDharPassive3");
-                                    explainedNumber.Add(choice.GetPassiveValue());
-                                }
-                            }
-                            
-                            
-                        }
-                    }
-                        
-                    
-                  
-                    return explainedNumber.ResultNumber;
-                }
+                TORAbilityModel  model = Campaign.Current.Models.GetAbilityModel();
+                return model.GetMaximumWindsOfMagic(this.BaseCharacter);
             }
         }
+
         public float WindsOfMagicRechargeRate
         {
             get
@@ -174,16 +110,9 @@ namespace TOR_Core.Extensions.ExtendedInfoSystem
                 if (!(Game.Current.GameType is Campaign)) return 0.2f;
                 else
                 {
-                    if (BaseCharacter.HeroObject != null && BaseCharacter.HeroObject != Hero.MainHero && BaseCharacter.HeroObject.Occupation == Occupation.Lord && BaseCharacter.HeroObject.IsSpellCaster()) return 2f;
-                    ExplainedNumber explainedNumber = new ExplainedNumber(1f, false, null);
-                    SkillHelper.AddSkillBonusForCharacter(TORSkills.SpellCraft, TORSkillEffects.WindsRechargeRate, BaseCharacter, ref explainedNumber);
-
-                    if ( BaseCharacter.HeroObject != null&& BaseCharacter.HeroObject.PartyBelongedTo!=null&&  BaseCharacter.HeroObject.PartyBelongedTo.IsMainParty )
-                    {
-                        CareerHelper.ApplyBasicCareerPassives(BaseCharacter.HeroObject, ref explainedNumber, PassiveEffectType.WindsRegeneration, false);
-                    }
-                    
-                    return explainedNumber.ResultNumber;
+                    if (!(Game.Current.GameType is Campaign)) return 50;
+                    TORAbilityModel  model = Campaign.Current.Models.GetAbilityModel();
+                    return model.GetWindsRechargeRate(this.BaseCharacter);
                 }
             }
         }
